# OOCL.Client/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "OOCL.Api",
		"version": "1.0"
	},
	"paths": {
...
		"/api/Image/images": {
			"get": {
				"tags": [
					"Image"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ImageObjInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ImageObjInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ImageObjInfo"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Image/images/{guid}/info": {
			"get": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Image/images/{guid}/remove": {
			"delete": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "boolean"
								}
							},
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							},
							"text/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Image/images/empty/{width}/{height}": {
			"post": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "width",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 1920
						}
					},
					{
						"name": "height",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 1080
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Image/images/upload": {
			"post": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "copyGuid",
						"in": "query",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							},
							"encoding": {
								"file": {
									"style": "form"
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Image/images/{guid}/download/{format}": {
			"get": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "format",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "bmp"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							},
							"application/json": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							},
							"text/json": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Image/images/{guid}/image64/{format}": {
			"get": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "format",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "bmp"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ImageData"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageData"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ImageData": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"readOnly": true
					},
					"name": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"filepath": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"meta": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"onHost": {
						"type": "boolean",
						"readOnly": true
					},
					"sizeMb": {
						"type": "number",
						"format": "float",
						"readOnly": true
					},
					"pointerHex": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataType": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataStructure": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"width": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"height": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"base64Format": {
						"type": "string",
						"nullable": true
					},
					"base64Image": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					}
				},
				"additionalProperties": false
			},
			"ImageObjInfo": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"readOnly": true
					},
					"name": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"filepath": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"meta": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"onHost": {
						"type": "boolean",
						"readOnly": true
					},
					"sizeMb": {
						"type": "number",
						"format": "float",
						"readOnly": true
					},
					"pointerHex": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataType": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataStructure": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"width": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"height": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"channels": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"bitdepth": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"lastLoadingTime": {
						"type": "number",
						"format": "double"
					},
					"lastProcessingTime": {
						"type": "number",
						"format": "double"
					},
					"errorMessage": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": false
			},
...
			"ProblemDetails": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "integer",
						"format": "int32",
						"nullable": true
					},
					"detail": {
						"type": "string",
						"nullable": true
					},
					"instance": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": {}
			}
		}
	}
}